requirements.txt
alembic==1.0.2
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.14
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
